Remove Agent ID (AID) If Building A Primary Image
Although the role includes an option to remove the AID, this is performed before the service tasks. This causes the sensor to create an AID when the sensor starts.

An additional fact has been added to set the state of the service to stopped when the variable falcon_remove_agent_id is true.

As a catch all, there is also a regedit to remove the AID keys on Windows install when falcon_remove_agent_id is true.

When the sensor is installed on a end point that isn't a primary image, the service is started by the service task.