Added new version of eventstream source
This version is similar except it doesn't use the task/consumer
async methods. The biggest difference here is that we don't
need to create a new queue to process events since our streaming
url is not an async operation. Instead, we yield an AsyncGenerate dict
object back to the queue to process.

The biggest thing here is that our producer (event streams firehose) can
overwhelm the queue, so we need to ensure we provide our queue with a
delay (min of 0.1 seems to work great) to ensure we get all events. This
is also now adjustable as an arg.