merge devel changes into main (#511)
* fix(kernel_support_info): adding new module that supports querying kernel compat info

* fix: update doc fragment to separate sort option

* fix: fixes ansible-test found issues

* chore(kernel_support_info): add changelog fragment

* feat: add new sensor update builds info module

Closes #454

This PR introduces the ability to query sensor update policy build
versions. This can be useful when determining either the latest sensor
versions, or the version associated with an auto-tagged build such as
N-1.

* chore: minor updates for galaxy and gitignore

* chore: adding changelog fragment for new module

Also moved the kernel_support_info fragment as for some reason it was in
the wrong directory.

* feat: new host_ids lookup plugin

This lookup plugin allows you to fetch a list of host ids based on a
search criteria. This is very useful to use for other host modules in
the collection. Fixes #502

* chore: add chagelog + update readme

* chore: fix linting issues

* feat: add new host_info module

Closes #479
This new module provides details for one or more hosts.

* chore: fix return docstring due to ansible-test issues

* chore: add changelog + update readme

* chore: add description for using lookup plugin

* docs: update description to include ref to new lookup

* docs: update lookup descriptions (#506)

* chore: update verbiage for better usage of plugin

Get rid of explicitly telling someone to escape double quotes and
instead provide them with options of how one can do this with/without
escaping double quotes.

* chore: update examples showing better way to use lookup plugin

* chore: update filter var name

* feat: add new lookup plugin to fetch maintenance token (#507)

* feat: add new lookup plugin to fetch maintenance token

Closes #490

This lookup plugin allows you to fetch a maintenance token for your
host, or a bulk token.

* add bulk option to retrieve bulk token

* minor updates

* Minor updates for 4.4.0 prep (#510)

* chore(build): add docs folder back to galaxy.yml

Fixes #508

* chore: aap 2.3 EOS changes

* chore(ci): update core version of ansible-test

* chore(docs): update readme to reflect new ansible core reqs

---------

Co-authored-by: Gabe Alford <redhatrises@gmail.com>