fix(sensor_download_info): use v2 override for new endpoint (#520)
* fix(sensor_download_info): use v2 override for new endpoint

To maintain backwards compatability with FalconPy, we take advantage of
the override feature that allows us to use the new
sensors/combined/installers/v2 endpoint. This means we can now pass in
architectures as a filter.

This also updates the limit to ensure users know the default limit is
set to 100.

* chore: minor wording changes to remove need for offset

This particular endpoint doesn't provide an offset which is strange
since the default limit is 100 and if you have more than that you
wouldn't know without passing in a larger limit.