fix(falcon_configure): this task should only run when both are set
There was an issue when the cid options was set to "" for it be deleted,
and then it would get to this point and fail because the when clause was
incorrect. A simple truthy test should provide us what we are looking
for. Also added options.cid to this since it makes sense we would want
to ensure both are set to use this task.