fix(falcon_install): add default sensor dl path to fix speed issues
Fixes #567

This PR fixes potential speed issues when working with large amounts of
hosts because for each host, it used to create a temp directory and
download the sensor. This of course is not very efficient, thus by
specifying a default such as /tmp - this should speed up download times
as hosts with the same sensor will not have to re-download every time.