feat(sensor_download): adds lock files to prevent collision when downloading similar sensors (#569)
* fix(falcon_install): add default sensor dl path to fix speed issues

Fixes #567

This PR fixes potential speed issues when working with large amounts of
hosts because for each host, it used to create a temp directory and
download the sensor. This of course is not very efficient, thus by
specifying a default such as /tmp - this should speed up download times
as hosts with the same sensor will not have to re-download every time.

* fix(falcon_install): make sure temp paths are static

This fixes an issue that will cause operations like copying the sensor
to create a new temp directory on the target host everytime it runs. For
example, in scenarios where something happens before the sensor is
cleaned, then running the role again will cause a new directory and
another full copy operation instead of checking to see if the sensor is
already copied.

* fix: cleaning up local sensor download dest should run once

* feat(sensor_download): adds the ability to lock files to prevent collision

This feature allows multiple hosts downloading the same file to not
overstep on each other as was the case before. This route improves
overall download times as the module will only download a sensor at most
once (assuming uniqueness) while other systems wait.

* chore: add sane retry_interval of 5 seconds

TBD - consider making this an option in the future?

* fix(falcon_install): win_temp_directory does not have path anymore

The win_file module does not return a path attribute. Instead of using
path, we'll just the variable instead as this is the path.

* chore: add delay to sensor_download flock

To prevent thundering herd!

* chore: add changelog

* lint: ignore bandit warning

* fix(falcon_install): update windows to use falcon_windows_tmp_dir variable

* fix(falcon_install): use win_stat to expand %SYSTEMROOT% properly

windows sucks... nuff said :(

* chore: changelog updates

* chore: add notes to sensor_download docs

Explain that a 0-byte lock file is created and can be safely deleted

* chore: revert back module arg 'name' changes

The refactoring was not needed and this makes more sense from an
organizational perspective.