fix(falcon_configure): control aid generation wait time and logic
Fixes #583

This PR should address concerns about potential timeouts or "misfires"
around waiting for an AID to be generated on Linux systems (particularly
on Ubuntu 20.04). We've added two ways to help mitigate these issues:
1. Use variables to control the retries/delays. Default to 60 seconds
(6/10).
2. Use a block/rescue so that if the 1st attempt fails, we give it
another try before concluding that something is wrong. We fail with a
message that includes some troubleshooting steps.